\name{GEO BreastCancer}
\alias{BreastCancer}
\docType{data}
\title{Subset of GEO GSE10810 Breast Cancer Dataset}
\description{
This dataset consists of the first 10 rows of a case/control mRNA
microarray dataset on human breast cancer. The intent is to provide
a small dataset that helps a user quickly become familiar
with the capabilities of EntropyExplorer, while at the
same time using actual experimental data in a standard format.
The dataset is divided into files "BreastCancer_case_10.txt" and
"BreastCancer_control_10.txt" that contain 31 tumor and 27 control
samples, respectively. Some of the samples are paired. The full dataset
and its description are available from the Gene Expression
Omnibus at http://www.ncbi.nlm.nih.gov/geo/ [1], accession number 
GSE10810 [2].
The data were obtained using the Affymetrix Human Genome U133 Plus 2.0 Array,
and contain mRNA expression values for more than 18k probes.
}

\source{Gene Expression Omnibus}
\references{
[1] Edgar R, Domrachev M, Lash AE.
Gene Expression Omnibus: NCBI gene expression and hybridization array data repository
Nucleic Acids Res. 2002 Jan 1;30(1):207-10

[2] Pedraza V, Gomez-Capilla JA, Escaramis G, Gomez C et al.
Gene expression signatures in breast cancer distinguish phenotype
characteristics, histologic subtypes, and tumor invasiveness.
Cancer 2010 Jan 15;116(2):486-96. 
}

\keyword{breast cancer, gene expression, microarrays}

